+++ examples/gpt2_example.py
@@ -22,9 +22,10 @@ from levanter import callbacks
 from levanter.config import TrainerConfig
 from levanter.data.sharded import GlobalBatchDataset
 from levanter.data.text import CachedLMDatasetConfig, TokenSeqDataset
+from levanter.grad_accum import accumulate_gradients_sharded
 from levanter.jax_utils import global_key_array, parameter_count
 from levanter.logging import capture_time, log_time_to_wandb
-from levanter.modeling_utils import accumulate_gradients_sharded, cross_entropy_loss_and_log_normalizers
+from levanter.modeling_utils import cross_entropy_loss_and_log_normalizers
 from levanter.models.gpt2 import Gpt2Config, Gpt2LMHeadModel
 from levanter.trainer_hooks import StepInfo, TrainerHooks
 from py_utils import non_caching_cycle
